resolved warnings in academic year

+++ app/src/main/java/gr/aueb/softeng/team02/Grade.java
@@ -3,16 +3,16 @@ package gr.aueb.softeng.team02;
 import java.util.ArrayList;
 
 public class Grade {
-    private double grade;
+    private float grade;
 
-    public Grade(double grade) {
+    public Grade(float grade) {
         this.grade = grade;
     }
     public double getGrade() {
         return grade;
     }
 
-    public void setGrade(double grade) {
+    public void setGrade(float grade) {
         // TODO Check exceptions cases and throws exception classes
         if (grade < 0 || grade > 10) {
             throw new RuntimeException("Invalid grade");
